Minor logic refactor, move dismiss logic to component, update tests
