+++ js/src/components/analytics-overview-promo/promo-actions.test.js
@@ -3,11 +3,24 @@
  */
 import '@testing-library/jest-dom';
 import { render, screen, fireEvent } from '@testing-library/react';
+import { useDispatch } from '@wordpress/data';
 
 /**
  * Internal dependencies
  */
+import { PREFERENCES_STORE_NAMESPACE } from '~/constants';
+import { ANALYTICS_OVERVIEW_PROMO_DISMISSED_KEY } from './constants';
 import PromoActions from './promo-actions';
+
+jest.mock( '@wordpress/data', () => ( {
+	__esModule: true,
+	useDispatch: jest.fn(),
+} ) );
+
+jest.mock( '@wordpress/preferences', () => ( {
+	__esModule: true,
+	store: 'preferences',
+} ) );
 
 jest.mock( '@wordpress/components', () => ( {
 	Flex: ( { children } ) => <div>{ children }</div>,
@@ -33,10 +46,15 @@ jest.mock( '~/utils/urls', () => ( {
 } ) );
 
 describe( 'PromoActions', () => {
+	const setPreference = jest.fn();
+
+	beforeEach( () => {
+		jest.clearAllMocks();
+		useDispatch.mockReturnValue( { set: setPreference } );
+	} );
+
 	test( 'renders the not-ready CTA', () => {
-		render(
-			<PromoActions isGoogleAdsReady={ false } onDismiss={ jest.fn() } />
-		);
+		render( <PromoActions isGoogleAdsReady={ false } /> );
 
 		expect(
 			screen.getByRole( 'link', { name: 'Get started' } )
@@ -44,23 +62,22 @@ describe( 'PromoActions', () => {
 	} );
 
 	test( 'renders the ready CTA', () => {
-		render(
-			<PromoActions isGoogleAdsReady={ true } onDismiss={ jest.fn() } />
-		);
+		render( <PromoActions isGoogleAdsReady={ true } /> );
 
 		expect(
 			screen.getByRole( 'link', { name: 'Launch a campaign' } )
 		).toHaveAttribute( 'href', '/create-campaign' );
 	} );
 
-	test( 'calls onDismiss when the Dismiss button is clicked', () => {
-		const onDismiss = jest.fn();
-		render(
-			<PromoActions isGoogleAdsReady={ false } onDismiss={ onDismiss } />
-		);
+	test( 'persists dismissal when the Dismiss button is clicked', () => {
+		render( <PromoActions isGoogleAdsReady={ false } /> );
 
 		fireEvent.click( screen.getByRole( 'button', { name: 'Dismiss' } ) );
 
-		expect( onDismiss ).toHaveBeenCalled();
+		expect( setPreference ).toHaveBeenCalledWith(
+			PREFERENCES_STORE_NAMESPACE,
+			ANALYTICS_OVERVIEW_PROMO_DISMISSED_KEY,
+			true
+		);
 	} );
 } );
