+++ js/src/components/analytics-overview-promo/promo-text.js
@@ -3,6 +3,12 @@
  */
 import { __ } from '@wordpress/i18n';
 
+/**
+ * @param {Object} props
+ * @param {string} props.title       The promo title.
+ * @param {string} props.description The promo description.
+ * @return {JSX.Element} The title/description pair.
+ */
 const TitleDescription = ( { title, description } ) => (
 	<>
 		<h3 className="gla-analytics-overview-promo__title">{ title }</h3>
@@ -21,22 +27,41 @@ const TitleDescription = ( { title, description } ) => (
  * @return {JSX.Element|null} The title/description pair, or `null` when `metricsCase` isn't recognized.
  */
 const PromoText = ( { metricsCase, isGoogleAdsReady } ) => {
-	if ( metricsCase === 'revenue' && ! isGoogleAdsReady ) {
-		return (
-			<TitleDescription
-				title={ __(
-					'Sales a bit slow? Reach more shoppers with Google.',
-					'google-listings-and-ads'
-				) }
-				description={ __(
-					'Sync your catalog with Google and grow back your sales by reaching new shoppers right when they are searching to buy.',
-					'google-listings-and-ads'
-				) }
-			/>
-		);
+	if ( ! isGoogleAdsReady ) {
+		if ( metricsCase === 'revenue' ) {
+			return (
+				<TitleDescription
+					title={ __(
+						'Sales a bit slow? Reach more shoppers with Google.',
+						'google-listings-and-ads'
+					) }
+					description={ __(
+						'Sync your catalog with Google and grow back your sales by reaching new shoppers right when they are searching to buy.',
+						'google-listings-and-ads'
+					) }
+				/>
+			);
+		}
+
+		if ( metricsCase === 'products' ) {
+			return (
+				<TitleDescription
+					title={ __(
+						'Selling fewer items than usual? Reach more shoppers with Google.',
+						'google-listings-and-ads'
+					) }
+					description={ __(
+						'Sync your catalog with Google and sell more of your products by reaching new shoppers right when they are searching to buy.',
+						'google-listings-and-ads'
+					) }
+				/>
+			);
+		}
+
+		return null;
 	}
 
-	if ( metricsCase === 'revenue' && isGoogleAdsReady ) {
+	if ( metricsCase === 'revenue' ) {
 		return (
 			<TitleDescription
 				title={ __(
@@ -51,22 +76,7 @@ const PromoText = ( { metricsCase, isGoogleAdsReady } ) => {
 		);
 	}
 
-	if ( metricsCase === 'products' && ! isGoogleAdsReady ) {
-		return (
-			<TitleDescription
-				title={ __(
-					'Selling fewer items than usual? Reach more shoppers with Google.',
-					'google-listings-and-ads'
-				) }
-				description={ __(
-					'Sync your catalog with Google and sell more of your products by reaching new shoppers right when they are searching to buy.',
-					'google-listings-and-ads'
-				) }
-			/>
-		);
-	}
-
-	if ( metricsCase === 'products' && isGoogleAdsReady ) {
+	if ( metricsCase === 'products' ) {
 		return (
 			<TitleDescription
 				title={ __(
